requirements.txt
cachetools==4.2.4; python_version >= "3.5" and python_version < "4.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
certifi==2021.10.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
charset-normalizer==2.0.8; python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
google-api-core==2.2.2; python_version >= "3.6" and python_version < "3.11"
google-auth==2.3.3; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
google-cloud-bigquery==2.30.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-core==2.2.1; python_version >= "3.6" and python_version < "3.11"
google-cloud-tasks==2.7.1; python_version >= "3.6"
google-crc32c==1.3.0; python_version >= "3.6" and python_version < "3.11"
google-resumable-media==2.1.0; python_version >= "3.6" and python_version < "3.11"
googleapis-common-protos==1.53.0; python_version >= "3.6" and python_version < "3.11"
grpc-google-iam-v1==0.12.3; python_version >= "3.6"
grpcio-status==1.42.0; python_version >= "3.6" and python_version < "3.11"
grpcio==1.42.0; python_version >= "3.6" and python_version < "3.11"
idna==3.3; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version >= "3.6" and python_version < "3.11"
packaging==21.3; python_version >= "3.6" and python_version < "3.11"
proto-plus==1.19.8; python_version >= "3.6" and python_version < "3.11"
protobuf==3.19.1; python_version >= "3.6" and python_version < "3.11"
pyasn1-modules==0.2.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
pyasn1==0.4.8; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0") or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
pyparsing==3.0.6; python_version >= "3.6" and python_version < "3.11"
python-dateutil==2.8.2; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.3.0"
requests==2.26.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.6.0")
rsa==4.8; python_version >= "3.6" and python_version < "4" and (python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0")
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_version >= "3.6" and python_version < "3.11" and python_full_version >= "3.6.0"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.11" or python_full_version >= "3.6.0" and python_version < "3.11" and python_version >= "3.6"
